# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Accounts Payable Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List all vendors
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create a vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get vendor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "404":
          description: Not found
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}/intelligence:
    get:
      operationId: getVendorIntelligence
      tags: [vendors]
      summary: Get vendor intelligence summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor intelligence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorIntelligence"
...
  /memory:
    get:
      operationId: listMemoryEvents
      tags: [memory]
      summary: List memory events
      parameters:
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: eventType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of memory events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryEvent"
    post:
      operationId: createMemoryEvent
      tags: [memory]
      summary: Create memory event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryEvent"
  /memory/search:
    post:
      operationId: searchMemory
      tags: [memory]
      summary: Semantic search over memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemorySearchInput"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryEvent"
...
components:
  schemas:
...
    Vendor:
      type: object
      required: [id, name, status, trustScore, disputeRate, totalInvoices, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        status:
          type: string
        trustScore:
          type: number
        disputeRate:
          type: number
        totalInvoices:
          type: integer
        totalAmount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    VendorInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        category:
          type: string
        notes:
          type: string
    VendorUpdate:
      type: object
      properties:
        name:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        category:
          type: string
        status:
          type: string
        trustScore:
          type: number
        notes:
          type: string
    VendorIntelligence:
      type: object
      required: [vendorId, summary, riskFactors, paymentBehavior, disputeHistory, recommendations]
      properties:
        vendorId:
          type: integer
        summary:
          type: string
        riskFactors:
          type: array
          items:
            type: string
        paymentBehavior:
          type: string
        disputeHistory:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        recentExceptions:
          type: array
          items:
            $ref: "#/components/schemas/Exception"
...
    Exception:
      type: object
      required: [id, vendorId, invoiceId, type, description, severity, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        vendorName:
          type: ["string", "null"]
        invoiceId:
          type: integer
        invoiceNumber:
          type: ["string", "null"]
        type:
          type: string
        description:
          type: string
        severity:
          type: string
        resolved:
          type: boolean
        resolvedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    MemoryEvent:
      type: object
      required: [id, vendorId, eventType, content, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        vendorName:
          type: ["string", "null"]
        invoiceId:
          type: ["integer", "null"]
        eventType:
          type: string
        content:
          type: string
        importance:
          type: ["number", "null"]
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
    MemoryEventInput:
      type: object
      required: [vendorId, eventType, content]
      properties:
        vendorId:
          type: integer
        invoiceId:
          type: integer
        eventType:
          type: string
        content:
          type: string
        importance:
          type: number
        tags:
          type: string
    MemorySearchInput:
      type: object
      required: [query]
      properties:
        query:
          type: string
        vendorId:
          type: integer
        limit:
          type: integer
...